Refactor Scalar DB Server code
