+++ server/src/integration-test/java/com/scalar/db/server/DistributedTransactionServiceWithJdbcTransactionIntegrationTest.java
@@ -17,7 +17,6 @@ import com.scalar.db.io.DataType;
 import com.scalar.db.io.IntValue;
 import com.scalar.db.io.Key;
 import com.scalar.db.io.Value;
-import com.scalar.db.server.config.ServerConfig;
 import com.scalar.db.service.StorageFactory;
 import com.scalar.db.storage.rpc.GrpcConfig;
 import com.scalar.db.transaction.rpc.GrpcTransaction;
